requirements.txt
#chainlit==1.0.200
langchain
langchain_community
PyPDF2
chromadb
groq
langchain-groq
ollama
python-dotenv
streamlit
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
transformers
sentence-transformers==2.2.2
InstructorEmbedding
ollama
huggingface_hub==0.21.4